Moved many folders from topics/ to the new once_useful_constructs/
