+++ once_useful_constructs/arithmetic.py
