+++ once_useful_constructs/combinatorics.py
