+++ once_useful_constructs/graph_theory.py
