+++ once_useful_constructs/light.py
